ysf_rx.grc: Default to RTL-SDR
So far, the example flowgraph was designed to play a file recorded from
a HackRF one at 8 MHz. This file was never commited to git though as
it's fairly large. Since this now has some faint level of maturity,
the hardware source was activated and the file source was disabled.

Similarly, since users are more likely to use an RTL-SDR than a HackRF
One on this, the default sampling rate and decimation has been adapted
for the RTL-SDR. The Getting Started HTML page lists example values
for both the HackRF One as well as the RTL-SDR. This fixes #23.